Moved CSharpScript over
